Corrected Forward text
I think the button text for BUTTON_MULTY_FORWARD should be "Forward as attachment(s)". The button seems to work when 1 or multiple e-mail messages are selected and always results in a new message with the email(s) attached as attachment, not in the message body like when using a normal single message forward button. Perhaps the label "BUTTON_MULTY_FORWARD" should also be changed to BUTTON_FORWARD_ATTACH or something troughout the code...

Same for the Help text for Shift-F. Perhaps the label "LABEL_FORWARD_MULTIPLY" should also be changed to LABEL_FORWARD_ATTACH troughout the code...

Ralf.